requirements.txt
-i https://pypi.org/simple
aniso8601==6.0.0
attrs==19.1.0
click==7.0
flask-restplus==0.12.1
flask==1.0.3
itsdangerous==1.1.0
jinja2==2.10.1
jsonschema==3.0.1
markupsafe==1.1.1
pyrsistent==0.15.2
pytz==2019.1
six==1.12.0
werkzeug==0.15.4
gunicorn==20.1.0
flask_restx
panda
numpy
sklearn 
joblib
